+++ wgpu-info/src/texture.rs
@@ -158,6 +158,56 @@ fn test_compute_render_extent() {
             let _ = desc.compute_render_extent(0, None);
         }
     }
+
+    for format in [wgpu::TextureFormat::NV12, wgpu::TextureFormat::P010] {
+        let desc = wgpu::TextureDescriptor {
+            label: None,
+            size: wgpu::Extent3d {
+                width: 8,
+                height: 4,
+                depth_or_array_layers: 1,
+            },
+            mip_level_count: 2,
+            sample_count: 1,
+            dimension: wgpu::TextureDimension::D2,
+            format,
+            usage: wgpu::TextureUsages::empty(),
+            view_formats: &[],
+        };
+
+        assert_eq!(
+            desc.compute_render_extent(0, Some(0)),
+            wgpu::Extent3d {
+                width: 8,
+                height: 4,
+                depth_or_array_layers: 1,
+            }
+        );
+        assert_eq!(
+            desc.compute_render_extent(0, Some(1)),
+            wgpu::Extent3d {
+                width: 4,
+                height: 2,
+                depth_or_array_layers: 1,
+            }
+        );
+        assert_eq!(
+            desc.compute_render_extent(1, Some(0)),
+            wgpu::Extent3d {
+                width: 4,
+                height: 2,
+                depth_or_array_layers: 1,
+            }
+        );
+        assert_eq!(
+            desc.compute_render_extent(1, Some(1)),
+            wgpu::Extent3d {
+                width: 2,
+                height: 1,
+                depth_or_array_layers: 1,
+            }
+        );
+    }
 }
 
 pub fn max_texture_format_string_size() -> usize {